Strengthened exchange API.

+++ settlement_provider/tests/integration/apiTrades.ts
@@ -1,12 +1,18 @@
-import { promises as fsPromises } from "fs"
+import { exists as existsAsync, promises as fsPromises } from "fs"
+import { promisify } from "util"
 import mockedEnv, { RestoreFn } from "mocked-env"
 import { expect } from "chai"
 import { createMocks } from "node-mocks-http"
-import handleTraderId from "../../pages/api/trader/[id]"
+import { OrderInfo } from "../../src/orderInfo"
+import { IExchangeDb } from "../../src/exchangeDb"
+import exchangeDbFactory from "../../src/exchangeDbFactory"
 import handleTrades from "../../pages/api/trades"
+
+const exists = promisify(existsAsync)
 
 describe("/api/trades Integration Tests", () => {
     let dbPath: string
+    let exchangeDb: IExchangeDb
     let toRestore: RestoreFn
 
     beforeEach("mock env", async() => {
@@ -14,11 +20,14 @@ describe("/api/trades Integration Tests", () => {
         toRestore = mockedEnv({
             "EXCHANGE_DB_PATH": dbPath
         })
+        exchangeDb = await exchangeDbFactory()
     })
     
     afterEach("restore env", async() => {
         toRestore()
-        await fsPromises.unlink(dbPath)
+        if (await exists(dbPath)) {
+            await fsPromises.unlink(dbPath)
+        }
     })
     
     describe("GET", () => {
@@ -35,21 +44,12 @@ describe("/api/trades Integration Tests", () => {
         })
     
         it("returns the info on previously set info", async () => {
-            {
-                const { req, res } = createMocks({
-                    "method": "PUT",
-                    "query": {
-                        "id": "3",
-                    },
-                    "body": {
-                        "isBuy": false,
-                        "quantity": 12345,
-                        "token": "ACME",
-                        "price": 33,
-                    }
-                })
-                await handleTraderId(req, res)
-            }
+            await exchangeDb.setOrderInfo("3", new OrderInfo({
+                "isBuy": false,
+                "quantity": 12345,
+                "token": "ACME",
+                "price": 33,
+            } as unknown as JSON))
             const { req, res } = createMocks({
                 "method": "GET"
             })
@@ -67,36 +67,18 @@ describe("/api/trades Integration Tests", () => {
         })
     
         it("returns the info on previously set double info", async () => {
-            {
-                const { req, res } = createMocks({
-                    "method": "PUT",
-                    "query": {
-                        "id": "3",
-                    },
-                    "body": {
-                        "isBuy": false,
-                        "quantity": 12345,
-                        "token": "ACME",
-                        "price": 33,
-                    }
-                })
-                await handleTraderId(req, res)
-            }
-            {
-                const { req, res } = createMocks({
-                    "method": "PUT",
-                    "query": {
-                        "id": "2",
-                    },
-                    "body": {
-                        "isBuy": true,
-                        "quantity": 543,
-                        "token": "ACME",
-                        "price": 30,
-                    }
-                })
-                await handleTraderId(req, res)
-            }
+            await exchangeDb.setOrderInfo("3", new OrderInfo({
+                "isBuy": false,
+                "quantity": 12345,
+                "token": "ACME",
+                "price": 33,
+            } as unknown as JSON))
+            await exchangeDb.setOrderInfo("2", new OrderInfo({
+                "isBuy": true,
+                "quantity": 543,
+                "token": "ACME",
+                "price": 30,
+            } as unknown as JSON))
             const { req, res } = createMocks({
                 "method": "GET"
             })
